+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - Consider `@JsonProperty.name` override for methods
 - Look-up `"description"` for methods (if included) based on `@JsonProperty.description`
+- Consider `@JsonProperty(access = Access.READ_ONLY)` when determining whether a field/method should be marked as `readOnly`
+- Consider `@JsonProperty(access = Access.WRITE_ONLY)` when determining whether a field/method should be marked as `writeOnly`
 
 ## [4.20.0] - 2021-09-04
 ### `jsonschema-generator`